src/trafficgym/engine/pyproject.toml
[project]
name = "sumo-workbench-engine"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "grpcio>=1.60",
  "grpcio-tools>=1.60",
  "protobuf>=4.25",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"
